+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.9.23
+
+### Patch Changes
+
+-   Updated dependencies [ea0a942]
+    -   @sap-ux/project-access@1.32.6
+
 ## 0.9.22
 
 ### Patch Changes