+++ packages/project-input-validator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/project-input-validator
 
+## 0.6.28
+
+### Patch Changes
+
+-   Updated dependencies [ea0a942]
+    -   @sap-ux/project-access@1.32.6
+
 ## 0.6.27
 
 ### Patch Changes